register enums with tag lookup
Was building the name of the tag constant var name in generator, but that assumed the tag was registered in the same package as the enum was being registered.  Doing a look up of the tag instead, which should work regardless of whether the tag is registed in same package, or in another.