Consolidated parsing and formatting code
All the code for transcoding KMIP values to strings as been consolidated in a set of formatting functions, which take a map of names to binary values as an argument.

All the registration of names to values has been consolidated in a new Registry type, and all the hidden package constants containing those maps is now consolidated in a single, package level DefaultRegistry.  All the code that transcodes KMIP values to and from strings now uses that DefaultRegistry.  The intent for the future is to enhance Encoder and Decoder to make the Registry configurable.

This meant changes to kmipgen.

Error messages were also improved and normalized.

Some docs were added.

Encoder now allows golang strings to be marshaled to Integer or Enumeration, if the field is explicitly flagged, or if the corresponding Tag has registered enum values.  The string is parsed according the same rules as JSON values.

Added a “bitmask” struct tag, for explicitly marking a struct field as an Integer, similar to the “enum” struct tag.