Fix for Attribute struct
The AttributeName in Attributes should be set to the canonical name from the spec, not the normalized, camelCase name.  Change the enum registry to preserve both forms of the name, and added methods to get the canonical name for values.  Also changed the docs: where I referred to "canonical", it now says "normalized".