When unmarshaling a struct value to interface{}, set the entire TTLV as the value
otherwise, it is very difficult to later unmarshal the interface{} value into another struct, which is a common pattern for slowly unmarshaling request and response messages one level at a time.  Unmarshaling just the value section of the TTLV struct makes it look like a stream of unrelated TTLV values, rather than fields of a structure.